TrainingAPI/requirements.txt
sanic==21.9.1
sanic-cors==1.0.1
sanic-session==0.8.0
sanic-openapi==21.6.1
sanic-testing==22.6.0
sanic-redis==0.3.0
termcolor==1.1.0
python-dotenv==0.19.2
jsonschema==3.2.0
pymongo==3.11.4
click==8.0.3
PyJWT==2.3.0
bcrypt==4.0.0

